requirements.txt
# Core framework
strands-agent>=0.1.0
strands-computer-use>=0.1.0
strands-browser-use>=0.1.0
aiohttp>=3.8.0
pydantic>=2.0.0

# AWS integration
boto3>=1.34.0
awscli>=1.32.0

# Communication APIs
slack-sdk>=3.26.0
requests>=2.31.0
O365>=2.0.27

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
typer>=0.9.0
pyyaml>=6.0.1
jsonschema>=4.20.0